Move Base32 builders to ...encoding.base32 pkg + deprecate ...encoding.builders pkg classes/functions
