+++ library/encoding-base32/api/encoding-base32.api
@@ -113,6 +113,59 @@ public final class io/matthewnelson/encoding/base32/Base32$Hex$Config : io/matth
 	public synthetic fun <init> (ZBZZLkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
+public final class io/matthewnelson/encoding/base32/Base32CrockfordConfigBuilder {
+	public field encodeToLowercase Z
+	public field finalizeWhenFlushed Z
+	public field hyphenInterval B
+	public field isLenient Z
+	public fun <init> ()V
+	public fun <init> (Lio/matthewnelson/encoding/base32/Base32$Crockford$Config;)V
+	public final fun build ()Lio/matthewnelson/encoding/base32/Base32$Crockford$Config;
+	public final fun checkSymbol ()Ljava/lang/Character;
+	public final fun checkSymbol (Ljava/lang/Character;)Lio/matthewnelson/encoding/base32/Base32CrockfordConfigBuilder;
+	public final fun strict ()Lio/matthewnelson/encoding/base32/Base32CrockfordConfigBuilder;
+}
+
+public final class io/matthewnelson/encoding/base32/Base32DefaultConfigBuilder {
+	public field encodeToLowercase Z
+	public field isLenient Z
+	public field lineBreakInterval B
+	public field padEncoded Z
+	public fun <init> ()V
+	public fun <init> (Lio/matthewnelson/encoding/base32/Base32$Default$Config;)V
+	public final fun build ()Lio/matthewnelson/encoding/base32/Base32$Default$Config;
+	public final fun strict ()Lio/matthewnelson/encoding/base32/Base32DefaultConfigBuilder;
+}
+
+public final class io/matthewnelson/encoding/base32/Base32HexConfigBuilder {
+	public field encodeToLowercase Z
+	public field isLenient Z
+	public field lineBreakInterval B
+	public field padEncoded Z
+	public fun <init> ()V
+	public fun <init> (Lio/matthewnelson/encoding/base32/Base32$Hex$Config;)V
+	public final fun build ()Lio/matthewnelson/encoding/base32/Base32$Hex$Config;
+	public final fun strict ()Lio/matthewnelson/encoding/base32/Base32HexConfigBuilder;
+}
+
+public final class io/matthewnelson/encoding/base32/BuildersKt {
+	public static final fun Base32Crockford ()Lio/matthewnelson/encoding/base32/Base32$Crockford;
+	public static final fun Base32Crockford (Lio/matthewnelson/encoding/base32/Base32$Crockford$Config;Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Crockford;
+	public static final fun Base32Crockford (Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Crockford;
+	public static final fun Base32Crockford (Z)Lio/matthewnelson/encoding/base32/Base32$Crockford;
+	public static synthetic fun Base32Crockford$default (ZILjava/lang/Object;)Lio/matthewnelson/encoding/base32/Base32$Crockford;
+	public static final fun Base32Default ()Lio/matthewnelson/encoding/base32/Base32$Default;
+	public static final fun Base32Default (Lio/matthewnelson/encoding/base32/Base32$Default$Config;Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Default;
+	public static final fun Base32Default (Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Default;
+	public static final fun Base32Default (Z)Lio/matthewnelson/encoding/base32/Base32$Default;
+	public static synthetic fun Base32Default$default (ZILjava/lang/Object;)Lio/matthewnelson/encoding/base32/Base32$Default;
+	public static final fun Base32Hex ()Lio/matthewnelson/encoding/base32/Base32$Hex;
+	public static final fun Base32Hex (Lio/matthewnelson/encoding/base32/Base32$Hex$Config;Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Hex;
+	public static final fun Base32Hex (Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base32/Base32$Hex;
+	public static final fun Base32Hex (Z)Lio/matthewnelson/encoding/base32/Base32$Hex;
+	public static synthetic fun Base32Hex$default (ZILjava/lang/Object;)Lio/matthewnelson/encoding/base32/Base32$Hex;
+}
+
 public final class io/matthewnelson/encoding/builders/Base32CrockfordConfigBuilder {
 	public field encodeToLowercase Z
 	public field finalizeWhenFlushed Z
